+++ stattools.hpp
@@ -18,21 +18,21 @@ void normalize_freqs(uint32_t* freqs,uint32_t* cum_freqs, size_t size,uint32_t t
 	uint32_t cur_total = cum_freqs[size];
 	
 	// resample distribution based on cumulative freqs
-	for (int i = 1; i <= size; i++)
+	for (size_t i = 1; i <= size; i++)
 		cum_freqs[i] = ((uint64_t)target_total * cum_freqs[i])/cur_total;
 
 	// if we nuked any non-0 frequency symbol to 0, we need to steal
 	// the range to make the frequency nonzero from elsewhere.
 	//
 	// this is not at all optimal, i'm just doing the first thing that comes to mind.
-	for (int i=0; i < size; i++) {
+	for (size_t i=0; i < size; i++) {
 		if (freqs[i] && cum_freqs[i+1] == cum_freqs[i]) {
 			// symbol i was set to zero freq
 
 			// find best symbol to steal frequency from (try to steal from low-freq ones)
 			uint32_t best_freq = ~0u;
 			int best_steal = -1;
-			for (int j=0; j < size; j++) {
+			for (size_t j=0; j < size; j++) {
 				uint32_t freq = cum_freqs[j+1] - cum_freqs[j];
 				if (freq > 1 && freq < best_freq) {
 					best_freq = freq;
@@ -43,11 +43,11 @@ void normalize_freqs(uint32_t* freqs,uint32_t* cum_freqs, size_t size,uint32_t t
 
 			// and steal from it!
 			if (best_steal < i) {
-				for (int j = best_steal + 1; j <= i; j++)
+				for (size_t j = best_steal + 1; j <= i; j++)
 					cum_freqs[j]--;
 			} else {
 				assert(best_steal > i);
-				for (int j = i + 1; j <= best_steal; j++)
+				for (size_t j = i + 1; j <= best_steal; j++)
 					cum_freqs[j]++;
 			}
 		}
@@ -55,7 +55,7 @@ void normalize_freqs(uint32_t* freqs,uint32_t* cum_freqs, size_t size,uint32_t t
 
 	// calculate updated freqs and make sure we didn't screw anything up
 	assert(cum_freqs[0] == 0 && cum_freqs[size] == target_total);
-	for (int i=0; i < size; i++) {
+	for (size_t i=0; i < size; i++) {
 		if (freqs[i] == 0)
 			assert(cum_freqs[i+1] == cum_freqs[i]);
 		else
